Refactor to modular data-driven architecture
- Replace hardcoded customer classes with JSON data structure
- Move all brand data to organized data/{brand}/ directories
- Extract rendering logic to InsightsRenderer class
- Implement dynamic brand loading from filesystem
- Remove customers/ module in favor of data-driven approach
```
